backend/finance/utils.py
import requests
import uuid
import logging

logger = logging.getLogger(__name__)

class MoMoIntegration:
    """
    Pattern for Ugandan Mobile Money Integration (MTN/Airtel).
    Typically uses an aggregator like Beyonic, Flutterwave, or direct telco APIs.
    """
    
    def __init__(self, provider='MTN'):
        self.provider = provider
        # These would be in environment variables / Tenant-specific settings
        self.api_key = "MOMO_API_KEY"
        self.base_url = "https://sandbox.momodeveloper.mtn.com" if provider == 'MTN' else "https://openapi.airtel.africa"

    def initiate_collection(self, phone_number, amount, external_id):
        """
        Triggers a Push (STK Push) on the user's phone.
        """
        logger.info(f"Initiating {self.provider} collection for {phone_number} - UGX {amount}")
        
        # simulated payload for MTN MoMo Collection Request
        payload = {
            "amount": str(amount),
            "currency": "UGX",
            "externalId": str(external_id),
            "payer": {
                "partyIdType": "MSISDN",
                "partyId": phone_number
            },
            "payerMessage": "School Fees Payment",
            "payeeNote": "Payment received"
        }
        
        # In a real scenario:
        # response = requests.post(f"{self.base_url}/collection/v1_0/requesttopay", json=payload, headers=headers)
        # return response.status_code, response.json()
        
        return 202, {"status": "SUCCESS", "transaction_id": str(uuid.uuid4())}

    def check_status(self, transaction_id):
        """
        Polls the status of a transaction.
        """
        # simulated check
        return {"status": "SUCCESS", "amount": "100000", "financialTransactionId": "123456789"}
